+++ src/main/java/com/backend/labpoint/dto/resource/ResourceDTO.java
@@ -3,5 +3,5 @@ package com.backend.labpoint.dto.resource;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record ResourceDTO(@NotNull Integer id, @NotBlank String name) {
+public record ResourceDTO(@NotNull Long id, @NotBlank String name) {
 }
