+++ src/main/java/com/backend/labpoint/dto/subject/DeleteSubjectRequestDTO.java
@@ -5,5 +5,5 @@ import jakarta.validation.constraints.NotEmpty;
 import java.util.Set;
 
 public record DeleteSubjectRequestDTO(
-        @NotEmpty(message = "O campo subjectIds nao pode ser nulo ou vazio!") Set<Integer> subjectIds) {
+        @NotEmpty(message = "O campo subjectIds nao pode ser nulo ou vazio!") Set<Long> subjectIds) {
 }
